Strip local PHPUnit harness — tests run via homeboy test only
Removed: phpunit.xml.dist, tests/bootstrap.php, tests/stubs/,
tests/Unit/, composer phpunit dependency.

Kept: tests/WebhookTest.php (16 integration tests using WP_UnitTestCase).
Blocked on homeboy-modules#19 (OB bug) before homeboy test works.
