+++ composer.json
@@ -12,9 +12,6 @@
     "require": {
         "php": ">=8.1"
     },
-    "require-dev": {
-        "phpunit/phpunit": "^10.0"
-    },
     "autoload": {
         "psr-4": {
             "Spawn\\": "inc/"
