+++ tests/WebhookTest.php
@@ -2,6 +2,8 @@
 /**
  * Tests for Spawn\Webhook.
  *
+ * Runs via Homeboy WordPress module: `homeboy test spawn`
+ *
  * @package Spawn
  */
 
@@ -11,15 +13,15 @@ use Spawn\Webhook;
 /**
  * Webhook handler tests.
  *
- * Uses pre_http_request filter to mock all external HTTP calls.
- * Tests run against SQLite via Homeboy WordPress module.
+ * Uses pre_http_request filter to mock external HTTP calls.
+ * Runs against real WordPress via Homeboy's WP test suite.
  */
 class WebhookTest extends WP_UnitTestCase {
 
 	/**
 	 * HTTP requests captured during test.
 	 *
-	 * @var array{url: string, args: array}[]
+	 * @var array[]
 	 */
 	private array $http_requests = [];
 
@@ -31,31 +33,18 @@ class WebhookTest extends WP_UnitTestCase {
 	private mixed $next_http_response = null;
 
 	/**
-	 * Emails captured during test.
-	 *
-	 * @var array[]
-	 */
-	private array $sent_emails = [];
-
-	/**
 	 * Set up before each test.
 	 */
 	public function set_up(): void {
 		parent::set_up();
 
-		// Create Spawn tables.
 		Database::create_tables();
 
-		// Set required options.
 		update_option( 'admin_email', 'admin@example.com' );
 		update_option( 'spawn_provisioner_url', 'http://127.0.0.1:8420' );
 		update_option( 'spawn_provisioner_token', 'test-token' );
 
-		// Intercept all outgoing HTTP requests.
 		add_filter( 'pre_http_request', [ $this, 'mock_http_request' ], 10, 3 );
-
-		// Capture emails.
-		add_filter( 'wp_mail', [ $this, 'capture_email' ] );
 	}
 
 	/**
@@ -67,11 +56,8 @@ class WebhookTest extends WP_UnitTestCase {
 		$wpdb->query( "DELETE FROM {$wpdb->prefix}spawn_domains" );
 
 		remove_filter( 'pre_http_request', [ $this, 'mock_http_request' ], 10 );
-		remove_filter( 'wp_mail', [ $this, 'capture_email' ] );
-
 		$this->http_requests      = [];
 		$this->next_http_response = null;
-		$this->sent_emails        = [];
 
 		parent::tear_down();
 	}
@@ -90,7 +76,7 @@ class WebhookTest extends WP_UnitTestCase {
 			'args' => $args,
 		];
 
-		if ( $this->next_http_response !== null ) {
+		if ( null !== $this->next_http_response ) {
 			$resp                     = $this->next_http_response;
 			$this->next_http_response = null;
 			return $resp;
@@ -103,29 +89,18 @@ class WebhookTest extends WP_UnitTestCase {
 		];
 	}
 
-	/**
-	 * Capture outgoing email.
-	 *
-	 * @param array $args Email args.
-	 * @return array Same args (passthrough).
-	 */
-	public function capture_email( array $args ): array {
-		$this->sent_emails[] = $args;
-		return $args;
-	}
-
 	// -----------------------------------------------------------------------
 	// Helpers
 	// -----------------------------------------------------------------------
 
 	/**
-	 * Create a test customer in the database.
+	 * Create a test customer.
 	 *
 	 * @param array $overrides Field overrides.
 	 * @return int Customer ID.
 	 */
 	private function create_test_customer( array $overrides = [] ): int {
-		$data = array_merge(
+		return Database::create_customer( array_merge(
 			[
 				'email'               => 'test@example.com',
 				'domain'              => 'old.example.com',
@@ -135,9 +110,7 @@ class WebhookTest extends WP_UnitTestCase {
 				'stripe_subscription' => 'sub_test',
 			],
 			$overrides
-		);
-
-		return Database::create_customer( $data );
+		) );
 	}
 
 	/**
@@ -166,27 +139,20 @@ class WebhookTest extends WP_UnitTestCase {
 	// Subscription Checkout
 	// -----------------------------------------------------------------------
 
-	/**
-	 * Test successful subscription checkout creates customer and triggers provisioner.
-	 */
 	public function test_subscription_checkout_creates_customer(): void {
 		$session = $this->subscription_session();
 		Webhook::handle_checkout_completed( $session, (object) [] );
 
 		$customer = Database::get_customer_by_domain( 'mysite.example.com' );
-		$this->assertNotNull( $customer, 'Customer should be created' );
+		$this->assertNotNull( $customer );
 		$this->assertEquals( 'subscriber@example.com', $customer['email'] );
 		$this->assertEquals( 'provisioning', $customer['status'] );
 		$this->assertEquals( 'cus_sub', $customer['stripe_customer'] );
 
-		// Provisioner should have been called.
 		$prov_requests = array_filter( $this->http_requests, fn( $r ) => str_contains( $r['url'], '8420' ) );
-		$this->assertNotEmpty( $prov_requests, 'Provisioner should be triggered' );
+		$this->assertNotEmpty( $prov_requests );
 	}
 
-	/**
-	 * Test checkout with missing email does nothing.
-	 */
 	public function test_subscription_checkout_missing_email(): void {
 		$session = $this->subscription_session( [ 'customer_email' => '' ] );
 		Webhook::handle_checkout_completed( $session, (object) [] );
@@ -196,29 +162,19 @@ class WebhookTest extends WP_UnitTestCase {
 		$this->assertEquals( 0, (int) $count );
 	}
 
-	/**
-	 * Test duplicate email with active status is rejected.
-	 */
 	public function test_subscription_checkout_duplicate_active_email(): void {
 		$this->create_test_customer( [ 'email' => 'subscriber@example.com', 'domain' => 'existing.com' ] );
 
 		$session = $this->subscription_session();
 		Webhook::handle_checkout_completed( $session, (object) [] );
 
-		// Should still only have the original customer.
 		global $wpdb;
 		$count = $wpdb->get_var( "SELECT COUNT(*) FROM {$wpdb->prefix}spawn_customers" );
 		$this->assertEquals( 1, (int) $count );
 	}
 
-	/**
-	 * Test duplicate domain is rejected.
-	 */
 	public function test_subscription_checkout_duplicate_domain(): void {
-		$this->create_test_customer( [
-			'email'  => 'first@example.com',
-			'domain' => 'mysite.example.com',
-		] );
+		$this->create_test_customer( [ 'email' => 'first@example.com', 'domain' => 'mysite.example.com' ] );
 
 		$session = $this->subscription_session();
 		Webhook::handle_checkout_completed( $session, (object) [] );
@@ -228,9 +184,6 @@ class WebhookTest extends WP_UnitTestCase {
 		$this->assertEquals( 1, (int) $count );
 	}
 
-	/**
-	 * Test provisioner failure sets customer status to failed.
-	 */
 	public function test_subscription_checkout_provisioner_failure(): void {
 		$this->next_http_response = new WP_Error( 'http_failure', 'Connection refused' );
 
@@ -246,19 +199,14 @@ class WebhookTest extends WP_UnitTestCase {
 	// Non-Spawn webhook
 	// -----------------------------------------------------------------------
 
-	/**
-	 * Test non-spawn webhooks are ignored.
-	 */
 	public function test_non_spawn_webhook_ignored(): void {
-		$session = [
-			'customer_email' => 'other@example.com',
-			'metadata'       => [
-				'source' => 'sell-my-images',
-				'job_id' => '123',
+		Webhook::handle_checkout_completed(
+			[
+				'customer_email' => 'other@example.com',
+				'metadata'       => [ 'source' => 'sell-my-images', 'job_id' => '123' ],
 			],
-		];
-
-		Webhook::handle_checkout_completed( $session, (object) [] );
+			(object) []
+		);
 
 		global $wpdb;
 		$count = $wpdb->get_var( "SELECT COUNT(*) FROM {$wpdb->prefix}spawn_customers" );
@@ -269,9 +217,6 @@ class WebhookTest extends WP_UnitTestCase {
 	// Credit Purchase
 	// -----------------------------------------------------------------------
 
-	/**
-	 * Test credit purchase webhook adds credits.
-	 */
 	public function test_credit_purchase_adds_credits(): void {
 		$customer_id = $this->create_test_customer( [
 			'email'           => 'buyer@example.com',
@@ -280,19 +225,20 @@ class WebhookTest extends WP_UnitTestCase {
 			'credit_balance'  => 5.00,
 		] );
 
-		$session = [
-			'customer'       => 'cus_buyer',
-			'customer_email' => 'buyer@example.com',
-			'amount_total'   => 2000,
-			'metadata'       => [
-				'source'            => 'spawn',
-				'type'              => 'credit_purchase',
-				'spawn_customer_id' => (string) $customer_id,
-				'credits'           => '20.00',
+		Webhook::handle_checkout_completed(
+			[
+				'customer'       => 'cus_buyer',
+				'customer_email' => 'buyer@example.com',
+				'amount_total'   => 2000,
+				'metadata'       => [
+					'source'            => 'spawn',
+					'type'              => 'credit_purchase',
+					'spawn_customer_id' => (string) $customer_id,
+					'credits'           => '20.00',
+				],
 			],
-		];
-
-		Webhook::handle_checkout_completed( $session, (object) [] );
+			(object) []
+		);
 
 		$customer = Database::get_customer( $customer_id );
 		$this->assertEquals( 25.00, (float) $customer['credit_balance'] );
@@ -302,37 +248,48 @@ class WebhookTest extends WP_UnitTestCase {
 	// Invoice Lifecycle
 	// -----------------------------------------------------------------------
 
-	/**
-	 * Test payment failed updates customer status.
-	 */
 	public function test_payment_failed_updates_status(): void {
-		$this->create_test_customer( [
-			'stripe_subscription' => 'sub_failing',
-		] );
+		$this->create_test_customer( [ 'stripe_subscription' => 'sub_failing' ] );
 
-		Webhook::handle_payment_failed(
-			[ 'subscription' => 'sub_failing' ],
-			(object) []
-		);
+		Webhook::handle_payment_failed( [ 'subscription' => 'sub_failing' ], (object) [] );
 
 		$customer = Database::get_customer_by_subscription( 'sub_failing' );
 		$this->assertEquals( 'payment_failed', $customer['status'] );
 	}
 
-	/**
-	 * Test payment failed skipped for comped customer.
-	 */
 	public function test_payment_failed_skipped_for_comped(): void {
 		$this->create_test_customer( [
 			'stripe_subscription' => 'sub_comped',
 			'billing_type'        => 'comped',
 		] );
 
-		Webhook::handle_payment_failed(
-			[ 'subscription' => 'sub_comped' ],
-			(object) []
-		);
+		Webhook::handle_payment_failed( [ 'subscription' => 'sub_comped' ], (object) [] );
 
+		$customer = Database::get_customer_by_subscription( 'sub_comped' );
+		$this->assertEquals( 'active', $customer['status'] );
+	}
+
+	public function test_invoice_paid_activates_customer(): void {
+		$this->create_test_customer( [
+			'stripe_subscription' => 'sub_renew',
+			'status'              => 'payment_failed',
+		] );
+
+		Webhook::handle_invoice_paid( [ 'subscription' => 'sub_renew' ], (object) [] );
+
+		$customer = Database::get_customer_by_subscription( 'sub_renew' );
+		$this->assertEquals( 'active', $customer['status'] );
+	}
+
+	public function test_invoice_paid_skipped_for_comped(): void {
+		$this->create_test_customer( [
+			'stripe_subscription' => 'sub_comped',
+			'billing_type'        => 'comped',
+		] );
+
+		Webhook::handle_invoice_paid( [ 'subscription' => 'sub_comped' ], (object) [] );
+
+		// Status should remain unchanged (no update call).
 		$customer = Database::get_customer_by_subscription( 'sub_comped' );
 		$this->assertEquals( 'active', $customer['status'] );
 	}
@@ -341,57 +298,36 @@ class WebhookTest extends WP_UnitTestCase {
 	// Subscription Cancelled
 	// -----------------------------------------------------------------------
 
-	/**
-	 * Test subscription cancelled schedules deletion.
-	 */
 	public function test_subscription_cancelled_schedules_deletion(): void {
-		$this->create_test_customer( [
-			'stripe_subscription' => 'sub_cancel',
-		] );
+		$this->create_test_customer( [ 'stripe_subscription' => 'sub_cancel' ] );
 
-		Webhook::handle_subscription_cancelled(
-			[ 'id' => 'sub_cancel' ],
-			(object) []
-		);
+		Webhook::handle_subscription_cancelled( [ 'id' => 'sub_cancel' ], (object) [] );
 
 		$customer = Database::get_customer_by_subscription( 'sub_cancel' );
 		$this->assertEquals( 'cancelling', $customer['status'] );
 		$this->assertNotEmpty( $customer['scheduled_deletion_at'] );
 	}
 
-	/**
-	 * Test subscription cancelled skipped for comped customer.
-	 */
 	public function test_subscription_cancelled_skipped_for_comped(): void {
 		$this->create_test_customer( [
 			'stripe_subscription' => 'sub_comped_cancel',
 			'billing_type'        => 'comped',
 		] );
 
-		Webhook::handle_subscription_cancelled(
-			[ 'id' => 'sub_comped_cancel' ],
-			(object) []
-		);
+		Webhook::handle_subscription_cancelled( [ 'id' => 'sub_comped_cancel' ], (object) [] );
 
 		$customer = Database::get_customer_by_subscription( 'sub_comped_cancel' );
 		$this->assertEquals( 'active', $customer['status'] );
 	}
 
-	/**
-	 * Test already-cancelling customer is not re-processed.
-	 */
 	public function test_subscription_cancelled_already_cancelling(): void {
 		$this->create_test_customer( [
 			'stripe_subscription' => 'sub_already',
 			'status'              => 'cancelling',
 		] );
 
-		Webhook::handle_subscription_cancelled(
-			[ 'id' => 'sub_already' ],
-			(object) []
-		);
+		Webhook::handle_subscription_cancelled( [ 'id' => 'sub_already' ], (object) [] );
 
-		// Status should remain cancelling (not re-processed).
 		$customer = Database::get_customer_by_subscription( 'sub_already' );
 		$this->assertEquals( 'cancelling', $customer['status'] );
 	}
@@ -400,9 +336,6 @@ class WebhookTest extends WP_UnitTestCase {
 	// Provisioner Webhook
 	// -----------------------------------------------------------------------
 
-	/**
-	 * Test provisioner completion webhook.
-	 */
 	public function test_provisioner_webhook_completion(): void {
 		$this->create_test_customer( [
 			'domain' => 'newsite.example.com',
@@ -422,15 +355,12 @@ class WebhookTest extends WP_UnitTestCase {
 
 		$response = Webhook::handle_provisioner_webhook( $request );
 
-		$this->assertInstanceOf( \WP_REST_Response::class, $response );
+		$this->assertInstanceOf( WP_REST_Response::class, $response );
 		$data = $response->get_data();
 		$this->assertTrue( $data['received'] );
 		$this->assertTrue( $data['processed'] );
 	}
 
-	/**
-	 * Test provisioner webhook with unknown event.
-	 */
 	public function test_provisioner_webhook_unknown_event(): void {
 		$request = new WP_REST_Request( 'POST' );
 		$request->set_body( wp_json_encode( [ 'event' => 'something_weird' ] ) );
